Refactors AppStateHandler into interface implementation pattern
