+++ vector/src/main/java/im/vector/app/features/home/room/list/RoomListViewModel.kt
@@ -148,7 +148,7 @@ class RoomListViewModel @AssistedInject constructor(
     private val roomListSectionBuilder = RoomListSectionBuilder(
                 session,
                 stringProvider,
-                appStateHandler,
+                appStateHandlerImpl,
                 viewModelScope,
                 autoAcceptInvites,
                 {
