+++ vector/src/main/java/im/vector/app/features/spaces/SpaceMenuViewModel.kt
@@ -73,7 +73,7 @@ class SpaceMenuViewModel @AssistedInject constructor(
                 it.getOrNull()?.let {
                     if (it.membership == Membership.LEAVE) {
                         setState { copy(leavingState = Success(Unit)) }
-                        if (appStateHandler.safeActiveSpaceId() == initialState.spaceId) {
+                        if (appStateHandler.getSafeActiveSpaceId() == initialState.spaceId) {
                             // switch to home?
                             appStateHandler.setCurrentSpace(null, session)
                         }
